pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "onc"
version = "2.3.5"
description = "Ocean 3.0 API Python Client Library"
readme = "README.md"
authors = [
    { name = "ONC Data Team", email = "data@oceannetworks.ca" },
]
license = { file = 'LICENSE.txt' }
requires-python = ">=3.8"
dependencies = [
    "requests",
    'python-dateutil',
    "humanize",
]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: Apache Software License',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3',
    'Topic :: Scientific/Engineering',
]

[project.urls]
Homepage = "https://wiki.oceannetworks.ca/display/O2A/Oceans+3.0+API+Home"
Documentation = "https://wiki.oceannetworks.ca/display/O2A/Client+Libraries"
Source = "https://github.com/OceanNetworksCanada/api-python-client"

[project.optional-dependencies]
dev = [
    "ipykernel",
    "python-dotenv",
    "tox",
    "black",
    "isort",
    "pytest",
    "pytest-cov",
    "ruff",
]

[tool.isort]
profile = "black"
src_paths = ["onc", "tests"]

[tool.ruff]
# Reference for the rules https://docs.astral.sh/ruff/rules/
select = [
    "E", "W", # pycodestyle
    "F", # pyflakes
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "SIM", # flake8-simplify
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
addopts = [
    "--import-mode=importlib",
]
